+++ tests-manual/test-plugins.sh
@@ -1,2 +1,0 @@
-PYTHONPATH=/ingest-validation-tools
-python -m tests-manual.update_test_data -t "examples/plugin-tests" -g "" --manual_test