+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Fix shared upload check for unreferenced files
 - Fix shared upload error messaging to display correct paths
 - Update CosMX Transcriptomics directory schema
+- Moved Validator and validation_class_iter to ingest-validation-tests
+- Deprecated cleanup examples
 
 ## v0.0.39
 - Create Object-by-Analyte EPIC directory schema